requirements.txt
fastapi
uvicorn
line-bot-sdk
langchain-groq
langchain-community
langchain-core
duckduckgo-search
gspread
google-auth
nest_asyncio